% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity.R
\name{dim_variety}
\alias{dim_variety}
\title{Variety or Richeness}
\usage{
dim_variety(data, sort = TRUE, entity_col = FALSE)
}
\arguments{
\item{data}{A numeric matrix with: entities as rows, categories as columns and cells as value of abundance. It could also be a matrix with categories in rows and entities in columns, but in that case, the paramenter "entity_col" should be set to TRUE. The matrix must include proper names for rows and columns. The parameter data, can also be a dataframe with three columns in this order: entities, categories, value of abundance.}

\item{sort}{Indicates if results should be ordered or not. Define it to FALSE to avoid ordering.}

\item{entity_col}{Entities are in columns. The analysis assumes that in the data matrix, the entities are in rows, but, if in the data matrix the entities are in the columns and the categories in the rows, then, the parameter entity_col, should be set to TRUE. Default is FALSE}
}
\value{
A dataframe with values of variety for each entity.
}
\description{
It computes the variety (number of distinct types) or simple diversity of an entity. It is also know as richeness.
}
\examples{
dim_variety(data=pantheon)
dim_variety(data=pantheon, sort=FALSE)
}

