% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity.R
\name{disparity}
\alias{disparity}
\title{A procedure to analyze a disparity matrix}
\usage{
disparity(X, method = "all", agg_type = "row")
}
\arguments{
\item{method}{List of available disparity matrix analysis methods: "avg", "sum", "all". The default for method is "all".}

\item{agg_type}{aggregation type for disparity analysis. The analysis is conducted per row but it can also be performed by column setting agg_type = "col". Default is "row".}

\item{data}{A disparity matrix to be analyzed}
}
\value{
A data frame with disparity measures as columns for each object of data
}
\description{
It takes a disparity matrix to get average or sums of disparities
}
\examples{
Xdis <- dist_mat(data, method="euclidean")
disp <- disparity(Xdis)
disp <- disparity(Xdis, method="avg", agg_type='col')
}

