% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity.R
\name{dist_mat}
\alias{dist_mat}
\title{A procedure to create a disparity matrix from a data frame or a matrix}
\usage{
dist_mat(data, method = NULL, agg_type = NULL)
}
\arguments{
\item{data}{A data frame or matrix of objects x categories}

\item{method}{List of available disparity methods: "cosine", "jaccard", "euclidean". The default for method is cosine.}

\item{agg_type}{aggregation type for pairwise disparity measure. The analysis is conducted per row but it can also be performed by column setting agg_type = "col". Default is row.}
}
\value{
A distance matrix
}
\description{
It takes a data frame or a matrix to create a disparity matrix
}
\examples{
Xdis <- dist_mat(data)
Xdis <- dist_mat(data, method="jaccard", agg_type='col')
}

