% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity.R
\name{distances}
\alias{distances}
\title{A procedure to create a disparity matrix from a data frame or a matrix}
\usage{
distances(data, method = "euclidean", entity_col = FALSE)
}
\arguments{
\item{data}{A data frame or matrix of objects x categories}

\item{method}{List of available disparity methods: "cosine", "jaccard", "euclidean". The default for method is cosine.}

\item{entity_col}{The entities are located in rows but, if they are located in columns, then entity_col should be set to TRUE. Default is FALSE.}
}
\value{
A distance matrix
}
\description{
It takes a data frame or a matrix to create a disparity matrix
}
\examples{
Xdis <- distances(pantheon)
Xdis <- distances(pantheon, method="jaccard", entity_col=TRUE)
Xdis <- distances(pantheon, method="cosine", entity_col=TRUE)
}

