% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity.R
\name{diver.bio}
\alias{diver.bio}
\title{Most common measures used in Ecology to analyze biodiversity}
\usage{
diver.bio(data, entity_col = FALSE)
}
\arguments{
\item{data}{A numeric matrix with: entities as rows, categories as columns and cells as value of abundance. It could also be a matrix with categories in rows and entities in columns, but in that case, the paramenter "entity_col" should be set to TRUE. The matrix must include proper names for rows and columns. The parameter data, can also be a dataframe with three columns in this order: entities, categories, value of abundance.}

\item{entity_col}{entities are in columns. The analysis assumes that in the data matrix, the entities are in rows, but, if in the data matrix the entities are in the columns and the categories in the rows, then, the parameter entity_col, should be set to TRUE. Default is FALSE}
}
\value{
a data frame with measures often used to measure diversity of an ecosystem
}
\description{
A procedure to compute the most common measures used to analyze the biodiversity of a ecoentity, such as Berger-Parker, Entropy and Simpson with their variations.
}
\examples{
str(geese)
diver.bio(geese)
}

