% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity.R
\name{diver.disparity}
\alias{diver.disparity}
\title{A procedure to compute the sum and average of disparities of entitys}
\usage{
diver.disparity(data, method = "cosine", entity_col = FALSE)
}
\arguments{
\item{data}{A numeric matrix with: entities as rows, categories as columns and cells as value of abundance. It could also be a matrix with categories in rows and entities in columns, but in that case, the paramenter "entity_col" should be set to TRUE. The matrix must include proper names for rows and columns. The parameter data, can also be a dataframe with three columns in this order: entities, categories, value of abundance.}

\item{method}{a distance measure available in proxy package.}

\item{entity_col}{entities are in columns. The analysis assumes that in the data matrix, the entities are in rows, but, if in the data matrix the entities are in the columns and the categories in the rows, then, the parameter entity_col, should be set to TRUE. Default is FALSE}
}
\value{
A data frame with resulting disparity measures of each entity in the dataset. Sum of disparities and average of disparities are computed.
}
\description{
It takes data of abundance of categories in entities and computes the sum and average of disparities between categories PRESENT in the entity
}
\examples{
diver.disparity(pantheon)
}

