% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity.R
\name{diver.ubiquity}
\alias{diver.ubiquity}
\title{Ubiquity of categories across entities}
\usage{
diver.ubiquity(data, entity_col = FALSE)
}
\arguments{
\item{data}{A numeric matrix with: entities as rows, categories as columns and cells as value of abundance. It could also be a matrix with categories in rows and entities in columns, but in that case, the paramenter "entity_col" should be set to TRUE. The matrix must include proper names for rows and columns. The parameter data, can also be a dataframe with three columns in this order: entities, categories, value of abundance.}

\item{entity_col}{entities are in columns. The analysis assumes that in the data matrix, the entities are in rows, but, if in the data matrix the entities are in the columns and the categories in the rows, then, the parameter entity_col, should be set to TRUE. Default is FALSE}
}
\value{
a dataframe with values of frequency per category. Decreasing order
}
\description{
It computes the ubiquity or the rearnes of the categories
}
\examples{
ub <- diver.ubiquity(data=pantheon)
ub
}

