% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{diverse}
\alias{diverse}
\alias{diverse-package}
\title{diverse: A package to compute diversity measures.}
\description{
The package \pkg{diverse} allows the user to compute the most common measures of diversity.
}
\details{
The main function of \pkg{diverse} package is
\code{\link{diversity}}. Other functions are useful according to user
needs, as \code{\link{dim_variety}} to compute simple variety-richness or
\code{\link{dim_balance}} to compute measures related to balance.
}

