% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geese}
\alias{geese}
\title{Geese data}
\format{A matrix with variables:
\describe{
\item{Columns}{Year of observation}
\item{Rows}{Specie}
}}
\source{
\url{https://www.sovon.nl/en}

\url{http://www.compass-project.eu/applets/3/index_EN.html}
}
\usage{
geese
}
\description{
Matrix of species of geese.
Dataset includes number of species of 4 geese observed by year in the Netherlands.
Data was curated by the Dutch bird protection organisation, Sovon.
}
\examples{
str(geese)
summary(geese)
}
\keyword{datasets}

