% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity.R
\name{get_data}
\alias{get_data}
\title{Get Data}
\usage{
get_data(data, entity_col = FALSE)
}
\arguments{
\item{data}{Data to be processed as dataframe or as matrix.}

\item{entity_col}{TRUE if column analysis is needed.}
}
\description{
It takes data as dataframe (edges) or as matrix (table) to be exported in proper form to be used by the diversity function.
}

