% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity.R
\name{plot.pheatmap}
\alias{plot.pheatmap}
\title{A procedure to plot the matrix of data as a pheatmap}
\usage{
\method{plot}{pheatmap}(data, title = NULL, fontsize = 14,
  fontsize_row = 7, color = c("blue", "yellow"))
}
\arguments{
\item{data}{A matrix of data to be ploted}

\item{fontsize}{The size of the font used in the plot.}

\item{fontsize_row}{The font size of the labels in the rows}
}
\description{
It takes a matrix of data and plots a pheatmap of that matrix
}
\examples{
pheatmap(data)
}

