% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity.R
\name{read_data}
\alias{read_data}
\title{A procedure to read data of a data file in formats csv, dta or spss}
\usage{
read_data(path, type = "csv", sep = ",", entity_col = FALSE)
}
\arguments{
\item{path}{A string representing the path to data file. If the data contained in the file is shaped as a matrix, the first column must include the names of the categories. If the data is shaped as edges list, it must contain three columns, which are: entity, category, value.}

\item{type}{It indicates the type of data to be read. This parameter facilitate the input of diverse type of data files, such as spss or stata. Possible options are the names of the mentioned softwares. The default value is csv.}

\item{sep}{Separator field used in the file to separate columns, if it is a CSV file. Default value is comma.}

\item{entity_col}{Entities are in the columns. The analysis assumes that the entities are in the rows of the matrix. If the entities are in the columns and the categories in the rows, then the parameter "entity_col" has to be set to TRUE. The default value is FALSE.}
}
\value{
A data frame with three columns (entity, category, value).
}
\description{
This function reads a file with data shaped as a matrix or as edges list. Several types of formats are allowed.
}
\examples{
#reading an edges list or panel shape, source data must include three columns
path <-  path_to_panel_file <- system.file("extdata", "PantheonEdges.csv", package = "diver")
sep <- ','
data <- read_data(path)
#reading a table
path <-  path_to_panel_file <- system.file("extdata", "PantheonMatrix.csv", package = "diver")
sep <- ','
data <- read_data(path)
#reading a table which includes the entities in the columns
path <-  path_to_matrix_file <- system.file("extdata", "Geese.csv", package = "diver")
data <- read_data(path, entity_col=TRUE)
}

