% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity.R
\name{readEdges}
\alias{readEdges}
\title{A procedure to read a list of edges from a file}
\usage{
readEdges(path, sepr, we = TRUE)
}
\arguments{
\item{path}{A string representing the path to a file. Rows and columns are described by integers (Ex.: 1 2 1)}

\item{sepr}{Separator field used in the file to separate columns}

\item{we}{It indicates if the list of edges includes weights or not. Default is TRUE}
}
\value{
A matrix with objects as rows and categories as rows
}
\description{
It takes a file and creates a matrix for diversity analysis
}
\examples{
path <- "~/MyDiversity/data/toy.edges"
sepr <- ' '
we <- TRUE
X <- readEdges(path,sepr,we)
}

