% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity.R
\name{readMatrix}
\alias{readMatrix}
\title{A procedure to read a matrix from a file}
\usage{
readMatrix(path, cols, sepr)
}
\arguments{
\item{path}{A string representing the path to a file. Each row corresponds to a matrix row (Ex.: 1 2 0 1 1 0 represents a row a matrix with 6 cols)}

\item{cols}{Number of columns of the matrix}

\item{sepr}{Separator field used in the file to separate columns}
}
\value{
A matrix with objects as rows and categories as rows
}
\description{
It takes a file and creates a matrix for diversity analysis
}
\examples{
path <- "~/MyDiversity/data/toy.mat"
cols <- 6
sepr <- ' '
X <- readMatrix(path,cols,sepr)
}

