% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scidat}
\alias{scidat}
\title{Scidat dataset}
\format{A matrix with variables:
\describe{
\item{Columns}{Areas of Science according SCimago}
\item{Rows}{Name of the country}
}}
\source{
Raw data before the aggregation was queried from
\url{http://pantheon.media.mit.edu/} in 2014.
}
\usage{
scidat
}
\description{
A matrix of number of papers authored from countries.
Dataset includes number of papers authored
from 10 countries in 2013. Data was aggregated from SCimago.
}
\examples{
str(scidat)
summary(scidat)
scidat["United States",]
scidat[,"Chemistry"]
}
\keyword{dataset}

