% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity.R
\name{sim_dataset}
\alias{sim_dataset}
\title{A procedure to simulate datasets}
\usage{
sim_dataset(n_categ, category_prefix = "", entity_prefix = "",
  values = "log-normal", size = -1, mean = 0, sd = 1)
}
\arguments{
\item{n_categ}{a vector with number of categories for each entity. The number of entities to create is defined by the length of this vector.}

\item{category_prefix}{a prefix for the label of the category}

\item{entity_prefix}{a prefix for the label of the entity}

\item{size}{number of individuals. A number or a vector of numbers for each entity. Default value is 7 times variety.}

\item{mean}{parameter for normal or log-normal distribution. Default value is 0.}

\item{sd}{parameter for normal or log-normal distribution. Default value is 1.}

\item{type}{both, a distribution name or a vector of integers. The distribution corresponds to  the function of 'categories (species) abundance distribution' to simulate individuals that are aggregated in frequencies or values of abundance. Use 'log-normal' for log normal distribution or 'normal' for normal distribution.  In the second case, an integer or a vector of integers of possible values of abundance to be used randomly. Default value is 'log-normal'}
}
\value{
A data frame with three columns: entity, category and value of abundance.
}
\description{
Simulates a dataset with values of variety for each entity and possible values of abundance.
}
\examples{
sim_dataset(n_categ=50,  category_prefix='ctg', values=1) #equal value
sim_entity(n_categ=50,  category_prefix='ctg', values=sample(1:100, replace=TRUE)) #random numbers for values of abundance
sim_entity(n_categ=50,  category_prefix='ctg', values='log-normal') #equal value
}

