% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity.R
\name{to_entropart}
\alias{to_entropart}
\title{A procedure to create a meta community for Entropart package}
\usage{
to_entropart(data, weights = 1)
}
\arguments{
\item{data}{a dataframe used in diverse with three columns, entity, category and value of abundance.}

\item{weights}{a vector of weights for each entity (community)}
}
\value{
An object of type metacommunity to be used in entropart.
}
\description{
Transform a dataframe used in diverse to a metacommunity used in Entropart.
}
\examples{
to_entropart(sim_dataset(c(1,2)))
}

