% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity.R
\name{u_distances}
\alias{u_distances}
\title{A procedure to create a disparity matrix from a data frame or a matrix}
\usage{
u_distances(data, method = "euclidean", entity_col = FALSE)
}
\arguments{
\item{data}{A numeric matrix with: entities as rows, categories as columns and cells as value of abundance. It could also be a matrix with categories in rows and entities in columns, but in that case, the paramenter "entity_col" should be set to TRUE. The matrix must include proper names for rows and columns. The parameter data, can also be a dataframe with three columns in this order: entities, categories, value of abundance.}

\item{method}{A distance or disimilarity method available in "proxy" package. A list of available methods can be queried by using the function \code{\link[proxy]{pr_DB}}. e.g. summary(pr_DB). If a similarity method is invoqued (as cosine) a proper transformation to disimilarity will be retrieved. Default is Euclidean.}

\item{entity_col}{Entities are in columns. The analysis assumes that in the data matrix, the entities are in rows, but, if in the data matrix the entities are in the columns and the categories in the rows, then, the parameter entity_col, should be set to TRUE. Default is FALSE}
}
\value{
A distance or disimilarity matrix
}
\description{
It takes a data frame or a matrix to create a disparity matrix
}
\examples{
Xdis <- u_distances(pantheon)
Xdis <- u_distances(pantheon, method="jaccard", entity_col=TRUE)
Xdis <- u_distances(pantheon, method="cosine", entity_col=TRUE)
}

