% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity.R
\name{ubiquity}
\alias{ubiquity}
\title{Ubiquity}
\usage{
ubiquity(data, entity_col = FALSE)
}
\arguments{
\item{data}{Data to be processed as dataframe or as matrix.}

\item{entity_col}{The entities are located in rows but, if they are located in columns, then entity_col should be set to TRUE. Default is FALSE}
}
\value{
a dataframe with values of frequency per category. Decreasing order
}
\description{
It computes the ubiquity or the rearnes of the categories
}
\examples{
ub <- ubiquity(data=pantheon)
ub
}

