% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity.R
\name{values}
\alias{values}
\title{Raw values of amount of elements in categories vs entities}
\usage{
values(data, category_row = FALSE, norm = NULL, filter = NULL,
  binary = FALSE)
}
\arguments{
\item{data}{A numeric matrix with entities \eqn{i} in the rows and categories \eqn{j} in the columns. Cells show the respective value (value of abundance) of entity \eqn{i} in the category \eqn{j}. It can also be a transpose of the previous matrix, that is, a matrix with categories in the rows and entities in the columns. Yet in that case, the parameter "category_row" has to be set to TRUE. The matrix must include names for the rows and the columns. The parameter "data", also accepts a dataframe with three columns in the following order: entity, category and value.}

\item{category_row}{A flag to indicate that categories are in the rows. The analysis assumes that the categories are in the columns of the matrix. If the categories are in the rows and the entities in the columns, then the parameter "category_row" has to be set to TRUE. The default value is FALSE.}

\item{norm}{A method to compute normalization. Possible values are 'p','proportions' 'rca' (Revealed Comparative Advantages)[Balassa 1986],'rca_norm' (RCA normalized between -1 and with 1. Values of RCA=1 equivalent to 0 in the normalization), 'ai' (Activity Index).}

\item{filter}{A thershold below which values are replaced with NA.}

\item{binary}{A boolean value to indicate if binarization is required. Values distinct than NA are replaced with 1.}
}
\value{
A matrix with raw values of number of elements in each category-entity.
}
\description{
It retrieves a matrix with raw values or with normalized values.
Also filter and binarization is available.
}
\details{
If the three parameters 'norm', 'filter' and 'binary' are used, then the same sequential order is used to make the calculation.
}
\examples{
val <- values(data=pantheon)
values(data = scidat)
values(data = scidat, norm='p')
values(data = scidat, norm='rca')
values(data = scidat, norm='rca', filter=1)
values(data = scidat, norm='rca', filter=1, binary=TRUE)
}
\references{
Balassa, B. (1986). Comparative advantage in manufactured goods: a reappraisal. The Review of Economics and Statistics, 315-319.
}

