% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity.R
\name{variety}
\alias{variety}
\title{Variety or Richeness}
\usage{
variety(data, sort = TRUE, entity_col = FALSE)
}
\arguments{
\item{data}{Data to be processed as dataframe or as matrix.}

\item{sort}{Indicates if results should be ordered or not. Define it to FALSE to avoid ordering.}

\item{entity_col}{The entities are located in rows but, if they are located in columns, then entity_col should be set to TRUE. Default is FALSE}
}
\value{
a dataframe with values of variety
}
\description{
It computes the variety (number of distinct types) or simple diversity of an entity. It is also know as richeness.
}
\examples{
variety(data=pantheon)
variety(data=pantheon, sort=FALSE)
}

