% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity.R
\name{writeCSV}
\alias{writeCSV}
\title{A procedure to write a diversity data frame to a csv file}
\usage{
writeCSV(frame, path, sepr)
}
\arguments{
\item{frame}{A data frame with diversity measures}

\item{path}{A string representing the path to a csv file}

\item{sepr}{Separator field used in the file to separate columns}
}
\value{
A csv file with objects as rows and diversity measures as cols
}
\description{
It takes a data frame and creates a csv file for diversity analysis
}
\examples{
writeCSV(data,"~/MyDiversity/data/data.csv",' ')
}

