% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity.R
\name{balance}
\alias{balance}
\title{Main measures of balance}
\usage{
balance(data, entity_col = FALSE)
}
\arguments{
\item{data}{A matrix of data with row and column names. Or a dataframe with three columns entity, category and value}

\item{entity_col}{The entities are located in rows but, if they are located in columns, then entity_col should be set to TRUE. Default is FALSE}
}
\value{
a dataset that includes main measures of balance
}
\description{
A procedure to compute several measures of the dimension balance of the diversity.
}
\examples{
str(pantheon)
}

