% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity.R
\name{disparity}
\alias{disparity}
\title{A procedure to compute the sum and average of disparities of entitys}
\usage{
disparity(data, method = "cosine", entity_col = FALSE)
}
\arguments{
\item{data}{A matrix or dataframe of entities, categories and values of presence}

\item{method}{a distance measure available in proxy package.}

\item{entity_col}{aggregation type for disparity analysis. The analysis is conducted per row but it can also be performed by column setting entity_col = TRUE. Default is FALSE.}
}
\value{
A data frame with disparity measures as columns for each entity of data. Sum of disparities and average of disparities are computed.
}
\description{
It takes data of presence-abundance of categories in entities and computes the sum and average of disparities between categories PRESENT in the entity
}
\examples{
disp <- disparity(pantheon)
}

