% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity.R
\name{disparity}
\alias{disparity}
\title{A procedure to compute the sum and average of disparities}
\usage{
disparity(data, method = "euclidean", category_row = FALSE)
}
\arguments{
\item{data}{A numeric matrix with entities \eqn{i} in the rows and categories \eqn{j} in the columns. Cells show the respective value (value of abundance) of entity \eqn{i} in the category \eqn{j}. It can also be a transpose of the previous matrix, that is, a matrix with categories in the rows and entities in the columns. Yet in that case, the parameter "category_row" has to be set to TRUE. The matrix must include names for the rows and the columns. The parameter "data", also accepts a dataframe with three columns in the following order: entity, category and value.}

\item{method}{A distance or dissimilarity method available in "proxy" package as for example "Euclidean", "Kullback" or "Canberra". This parameter also accepts a similarity method available in the "proxy" package, as for example: "cosine", "correlation" or "Jaccard" among others. In the latter case, a correspondent transformation to a dissimilarity measure will be retrieved. A list of available methods can be queried by using the function \code{\link[proxy]{pr_DB}}. e.g. summary(pr_DB). The default value is Euclidean distance.}

\item{category_row}{A flag to indicate that categories are in the rows. The analysis assumes that the categories are in the columns of the matrix. If the categories are in the rows and the entities in the columns, then the parameter "category_row" has to be set to TRUE. The default value is FALSE.}
}
\value{
A data frame with disparity measures for each entity in the dataset. Both the sum of disparities and the average of disparities are computed.
}
\description{
Computes the sum and the average of distances or disparities between the categories.
}
\examples{
data(pantheon)
disparity(data= pantheon)
disparity(data = pantheon, method='Canberra')
#For scientific publications
#Same disparities, since all countries authored all entities
disparity(scidat)
disparity(data= scidat, method='cosine')
#Creating differences by measuring Revealed Compartive Advantages
disparity(values(scidat, norm='rca', filter=1))
#Activity Index for scientometrics
disparity(values(scidat, norm='ai', filter=0), method='cosine')
#Using binarization of values and a binary metric for dissimilarities.
disparity(values(scidat, norm='ai', filter=0, binary=TRUE), method='jaccard')
}

