% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geese}
\alias{geese}
\title{Geese dataset}
\format{A matrix with variables:
\describe{
\item{Columns}{Year of observation}
\item{Rows}{Specie}
}}
\source{
\url{https://www.sovon.nl/en}

\url{http://www.compass-project.eu/applets/3/index_EN.html}
}
\usage{
geese
}
\description{
A matrix of species of geese.
Dataset includes the quantity of 4 species of geese observed by year in the Netherlands.
Data was curated by the Dutch bird protection organisation Sovon.
}
\examples{
data(geese)
str(geese)
summary(geese)
geese[,"2000"]
geese["Mute Swan",]
}
\keyword{dataset}

