% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity.R
\name{read.data}
\alias{read.data}
\title{A procedure to read data of a data file in formats csv, dta or spss}
\usage{
read.data(path, type = "csv", sep = ",", entity_col = FALSE)
}
\arguments{
\item{path}{A string representing the path to data file. If it is shaped as a matrix, the first column must include proper names of the categories. If it is shaped as edges list, it must contain three columns, which are entity, category, value.}

\item{type}{It indicates the type of data to be read. This parameter facilitate the input of diverse type of data files, such as spss or stata. Posible options are the names of the mentioned softwares. Default value is csv.}

\item{sep}{Separator field used in the file to separate columns, if it is a CSV file. Default value is comma.}
}
\value{
A data frame with three columns, even when the input file is shaped as a matrix.
}
\description{
It reads a file with data shaped as a matrix or as edges list. Several types of formats are allowed.
}
\examples{
#reading an edges list or panel shape, source data must include three columns
path <-  path_to_panel_file <- system.file("extdata", "PantheonEdges.csv", package = "diver")
sep <- ','
data <- read.data(path)
#reading a table
path <-  path_to_panel_file <- system.file("extdata", "PantheonMatrix.csv", package = "diver")
sep <- ','
data <- read.data(path)
#reading a table which includes the entities in the columns
path <-  path_to_matrix_file <- system.file("extdata", "Geese.csv", package = "diver")
data <- read.data(path, entity_col=TRUE)
}

