% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity.R
\name{to_entropart}
\alias{to_entropart}
\title{Transforms data to be used in Entropart package}
\usage{
to_entropart(data)
}
\arguments{
\item{data}{a dataframe used in diverse with three columns, entity, category and value of abundance.}
}
\value{
An object of type matrix of abundance to be used in entropart to create metacommunities.
}
\description{
Transform a dataframe used in diverse to values of abundance to be used in Entropart.
}
\examples{
ab <- to_entropart(sim_dataset(c(1,2)))
}

