% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity.R
\name{ubiquity}
\alias{ubiquity}
\title{Ubiquity}
\usage{
ubiquity(data)
}
\arguments{
\item{data}{Data to be processed as dataframe or as matrix.}

\item{data_agg}{Diferent of NULL if column analysis is needed.}
}
\value{
a dataframe with values of frequency per category. Decreasing order
}
\description{
It computes the ubiquity or the rearnes of the categories
}
\examples{
ub <- ubiquity(data=d)
}

